backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
aiomysql==0.2.0
pymysql==1.1.0
cryptography==42.0.0

# Redis
redis[hiredis]==5.0.1

# RabbitMQ
aio-pika==9.3.1

# MinIO Object Storage
minio==7.2.3

# OCR Engines
paddleocr==3.3.0
paddlepaddle==3.2.0
pytesseract==0.3.13  # Python wrapper for Tesseract 5.5
pdf2image==1.17.0

# PDF Processing
PyPDF2==3.0.1

# UmiOCR 通过 HTTP API 调用，无需额外依赖
# httpx 已在 Utilities 部分包含

# LLM Integration
agently>=4.0.6

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx>=0.28.1,<0.29.0
aiofiles==23.2.1
pillow==10.2.0
js2py==0.74  # JavaScript expression execution

# Logging and Monitoring
loguru==0.7.2

# Task Scheduling
apscheduler==3.10.4

# Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3

